240415 이효진 자재 단위 통일 중

+++ src/main/java/com/example/demo/materials/MaterialVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class MaterialVO {
 	private String matCode;
 	private String matName;
-	private Integer stockCnt;
+	private Double stockCnt;
 	private Integer safeStockCnt;
 	private String mainCompanyCode;
 	private Integer unitPrice;
